moved capactiy-table
