+++ src/Tickets/Seating/app/blockEditor/index.js
@@ -1,2 +1,3 @@
 import './store';
 import './filters.js';
+import './capacity-table';
